python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pybaseball>=2.2.7
pandas==2.1.4
numpy==1.25.2
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
openai==1.97.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
aiofiles==23.2.0
tenacity==8.2.3